Adjust files according to file list changes
