+++ lib/al/Library/LiveActor/LiveActorInfo.cpp
@@ -1,0 +1,7 @@
+#include "Library/LiveActor/LiveActorInfo.h"
+
+namespace al {
+
+SubActorInfo::SubActorInfo() = default;
+
+}
