+++ lib/al/Library/MapObj/SubActorLodMapParts.cpp
@@ -3,7 +3,7 @@
 #include "Library/LiveActor/ActorActionFunction.h"
 #include "Library/LiveActor/ActorInitUtil.h"
 #include "Library/LiveActor/ActorModelFunction.h"
-#include "Library/LiveActor/SubActorKeeper.h"
+#include "Library/LiveActor/LiveActorFunction.h"
 #include "Library/MapObj/SubActorLodExecutor.h"
 #include "Library/Placement/PlacementFunction.h"
