+++ src/Enemy/SenobiLeaf.cpp
@@ -9,7 +9,7 @@
 #include "Library/LiveActor/ActorInitUtil.h"
 #include "Library/LiveActor/ActorModelFunction.h"
 #include "Library/LiveActor/ActorPoseUtil.h"
-#include "Library/LiveActor/SubActorKeeper.h"
+#include "Library/LiveActor/LiveActorFunction.h"
 #include "Library/Math/MathUtil.h"
 
 SenobiLeaf::SenobiLeaf(const char* actorName) : al::LiveActor(actorName) {}
