+++ src/Scene/StageSceneStatePauseMenu.cpp
@@ -19,7 +19,7 @@
 #include "Library/LiveActor/ActorSceneInfo.h"
 #include "Library/LiveActor/LiveActor.h"
 #include "Library/LiveActor/LiveActorKit.h"
-#include "Library/LiveActor/SubActorKeeper.h"
+#include "Library/LiveActor/LiveActorFunction.h"
 #include "Library/Math/MathUtil.h"
 #include "Library/Message/MessageHolder.h"
 #include "Library/Nerve/NerveSetupUtil.h"
